requirements.txt
#TODO cleanup
anyio==3.6.2
#autopep8==2.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna==3.4
pycodestyle==2.10.0
pydantic==1.10.4
#PyJWT==2.6.0
python-decouple==3.7
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
